requirements.txt
alembic==0.9.5
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-Cors==3.0.3
Flask-HTTPAuth==3.2.3
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
passlib==1.7.1
pbr==3.1.1
psycopg2==2.7.3.1
python-dateutil==2.6.1
python-editor==1.0.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.14
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
urllib3==1.22
Werkzeug==0.12.2
